added services for crud for user, movie, theatre, screen and show

+++ src/main/java/com/bms/repositories/BookingRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingRepository extends CrudRepository<BookingEntity, Integer>{
+public interface BookingRepository extends CrudRepository<BookingEntity, Long>{
 }
