+++ src/main/java/com/bms/repositories/MovieRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovieRepository extends CrudRepository<MovieEntity, Integer>{
+public interface MovieRepository extends CrudRepository<MovieEntity, Long>{
 
 	@Query("SELECT m FROM MovieEntity m WHERE m.name=:name")
 	Iterable<MovieEntity> findMovieByName(@Param("name") String name);
