+++ src/main/java/com/bms/repositories/PaymentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentRepository extends CrudRepository<PaymentEntity, Integer>{
+public interface PaymentRepository extends CrudRepository<PaymentEntity, Long>{
 }
