+++ src/main/java/com/bms/repositories/SeatsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeatsRepository extends CrudRepository<SeatsEntity, Integer>{
+public interface SeatsRepository extends CrudRepository<SeatsEntity, Long>{
 }
